refactor(models): move shared contracts to lower-layer homes
Move role/execution shared contracts from execution/ layer to models/ and config/ layers to eliminate roles → execution import violations.

Changes:
- Create src/vibe3/models/worktree.py with WorktreeRequirement enum
- Create src/vibe3/models/session_types.py with SessionRole type alias
- Create src/vibe3/models/execution_request.py with ExecutionRequest and ExecutionLaunchResult dataclasses
- Create src/vibe3/config/role_gates.py with gate config constants
- Update execution/role_contracts.py to re-export from new locations
- Update execution/contracts.py to re-export DTOs from models/
- Update execution/session_service.py to re-export SessionRole from models/
- Update execution/__init__.py to import from new locations
- Update all roles/ imports to use models.* and config.role_gates
- Update execution-internal imports (coordinator, auto_scene_recovery, etc.)
- Update domain/handlers imports
- Update services/error_helpers.py TYPE_CHECKING import
- Update models/__init__.py to re-export new types

All existing imports continue to work through backward-compatible re-exports.

Refs: #1882
