+++ src/vibe3/execution/governance_sync_runner.py
@@ -15,9 +15,9 @@ from typer import echo
 from vibe3.agents.backends.codeagent import CodeagentBackend
 from vibe3.clients.store_context import get_store
 from vibe3.config.orchestra_settings import load_orchestra_config
-from vibe3.execution.contracts import ExecutionLaunchResult, ExecutionRequest
-from vibe3.execution.role_contracts import GOVERNANCE_GATE_CONFIG
+from vibe3.config.role_gates import GOVERNANCE_GATE_CONFIG
 from vibe3.execution.role_interfaces import GovernanceEventLogger, GovernanceFunctions
+from vibe3.models.execution_request import ExecutionLaunchResult, ExecutionRequest
 from vibe3.services.error_helpers import record_dispatch_failure_if_unexpected
