+++ src/vibe3/execution/session_service.py
@@ -1,7 +1,6 @@
 """Execution session resume helpers."""
 
 import re
-from typing import Literal
 
 from loguru import logger
 
@@ -9,10 +8,7 @@ from vibe3.clients.git_client import GitClient
 from vibe3.clients.sqlite_client import SQLiteClient
 from vibe3.environment.session_registry import SessionRegistryService
 from vibe3.exceptions import SystemError, UserError
-
-SessionRole = Literal[
-    "manager", "planner", "executor", "reviewer", "supervisor", "governance"
-]
+from vibe3.models.session_types import SessionRole
 
 _SESSION_ID_RE = re.compile(r"^[A-Za-z0-9_-]+$")
