+++ src/vibe3/models/__init__.py
@@ -10,9 +10,11 @@ from vibe3.models.change_source import (
 )
 from vibe3.models.coverage import CoverageReport, LayerCoverage
 from vibe3.models.dead_code import DeadCodeFinding, DeadCodeReport
+from vibe3.models.execution_request import ExecutionLaunchResult, ExecutionRequest
 from vibe3.models.inspection import CallNode, CommandInspection
 from vibe3.models.orchestra_config import OrchestraConfig
 from vibe3.models.prompt_meta import PromptContextMode
+from vibe3.models.session_types import SessionRole
 from vibe3.models.snapshot import (
     DependencyChange,
     DependencyEdge,
@@ -28,6 +30,7 @@ from vibe3.models.snapshot import (
     StructureSnapshot,
 )
 from vibe3.models.trace import ExecutionStep, TraceOutput
+from vibe3.models.worktree import WorktreeRequirement
 
 __all__: list[str] = [
     "BranchSource",
@@ -43,6 +46,8 @@ __all__: list[str] = [
     "DependencyEdge",
     "DiffSummary",
     "DiffWarning",
+    "ExecutionLaunchResult",
+    "ExecutionRequest",
     "ExecutionStep",
     "FileChange",
     "FileSnapshot",
@@ -53,9 +58,11 @@ __all__: list[str] = [
     "OrchestraConfig",
     "PRSource",
     "PromptContextMode",
+    "SessionRole",
     "StructureDiff",
     "StructureMetrics",
     "StructureSnapshot",
     "TraceOutput",
     "UncommittedSource",
+    "WorktreeRequirement",
 ]
