+++ src/vibe3/models/session_types.py
@@ -1,0 +1,7 @@
+"""Session role type definitions."""
+
+from typing import Literal
+
+SessionRole = Literal[
+    "manager", "planner", "executor", "reviewer", "supervisor", "governance"
+]
