+++ src/vibe3/roles/run_command.py
@@ -20,12 +20,12 @@ from vibe3.config.settings import VibeConfig
 from vibe3.exceptions import SkillNotAvailableError
 from vibe3.execution.codeagent_runner import CodeagentExecutionService
 from vibe3.execution.codeagent_support import build_self_invocation
-from vibe3.execution.contracts import ExecutionRequest
 from vibe3.execution.coordinator import ExecutionCoordinator
 from vibe3.execution.prompt_meta import build_prompt_meta
-from vibe3.execution.role_contracts import WorktreeRequirement
 from vibe3.execution.session_service import load_session_id
+from vibe3.models.execution_request import ExecutionRequest
 from vibe3.models.prompt_meta import PromptContextMode
+from vibe3.models.worktree import WorktreeRequirement
 from vibe3.roles.run_helpers import (
     publish_run_command_failure,
     publish_run_command_success,
