+++ src/vibe3/roles/supervisor.py
@@ -6,14 +6,14 @@ from collections.abc import Iterable
 from pathlib import Path
 from typing import Any
 
-from vibe3.domain.events.supervisor_apply import SupervisorIssueIdentified
-from vibe3.execution.contracts import ExecutionRequest
-from vibe3.execution.execution_role_policy import ExecutionRolePolicyService
-from vibe3.execution.issue_role_support import use_current_branch
-from vibe3.execution.role_contracts import (
+from vibe3.config.role_gates import (
     SUPERVISOR_APPLY_GATE_CONFIG,
     SUPERVISOR_IDENTIFY_GATE_CONFIG,
 )
+from vibe3.domain.events.supervisor_apply import SupervisorIssueIdentified
+from vibe3.execution.execution_role_policy import ExecutionRolePolicyService
+from vibe3.execution.issue_role_support import use_current_branch
+from vibe3.models.execution_request import ExecutionRequest
 from vibe3.models.orchestra_config import OrchestraConfig
 from vibe3.models.orchestration import IssueInfo
 from vibe3.roles.definitions import IssueRoleSyncSpec, RoleDefinition
